OKX_Straddle/requirements.txt
python-dotenv
web3
pydantic
asyncio
python-okx
python-telegram-bot